+++ src/etha/comm/__init__.py
@@ -1,9 +1,10 @@
 """Communication utilities."""
 
-from .chunk_ops import Transfer
+from .ir import Transfer
 from .get_m2m_map import get_m2m_map
 from .comm_methods import m2m_communicate, gather_broadcast_communicate
-from .get_chunk_ops import get_m2m_transfers, transfers_to_chunks, bind_tensors_to_chunks
+from .get_chunk_ops import transfers_to_chunks, bind_tensors_to_chunks
+from .get_transfers import get_m2m_transfers
 
 __all__ = [
     "get_m2m_map",
